Minor updates to remove warnings when running.
The csv library which this project uses directly and the ostruct library used by dependencies of this project are being moved out of the standard ruby library. So include these explicitly in the gemspec.

Also explicitly declare `Money.rounding_mode = BigDecimal::ROUND_HALF_EVEN` for the entire library since that will no longer be the default in future releases of the Money library.